% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross validation for selecting the thresholding parameter in CSNet}
\usage{
cross_validation(
  data_list,
  method_pars = list(methods_settings = list(var = "nnls", covar = "wls"), th = 0.01),
  th_func = "SCAD",
  n_th = 100,
  cv_fold = 2,
  n_splits = 10,
  ct_names = NULL,
  to_plot = F,
  seed = 1
)
}
\arguments{
\item{data_list}{a list of the matrix of cell type proportions P and the matrix
of bulk gene expression X.}

\item{method_pars}{parameters for runinng mom_ls(). Default to nnls and wls
for least squares estimation, and 0.01 for minimal weight.}

\item{th_func}{thresholding operator. Default to SCAD.}

\item{n_th}{number of threshold candidates. Default to 100.}

\item{cv_fold}{number of cross validation folds. Default to 2 for equal split.}

\item{n_splits}{number of 2-fold splits. Default to 10.}

\item{ct_names}{the name of cell types}

\item{to_plot}{whether to plot the cross validation results. Default to F.}

\item{seed}{random seed. Default to 1.}
}
\value{
A list of K cell-type-specific variance-covariance matrices.
}
\description{
This function implements cross validation to select the thresholding parameter
for the CSNet estimator.
}
