% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSNet.R
\name{CSNet}
\alias{CSNet}
\title{CSNet main function}
\usage{
CSNet(
  X,
  P,
  ls_methods = list(var = "nnls", covar = "wls"),
  weight_th = 0.01,
  plot_cv = F
)
}
\arguments{
\item{X}{matrix of bulk gene expression. n samples by p genes.}

\item{P}{matrix of cell type proportions. n samples by K cell types.}

\item{ls_methods}{least squares method to estimate variance and covariance.
Default to list(var='nnls', covar = 'wls'), i.e. non-negative least squares
for estimating variance and weighted least squares for estimating covariance.}

\item{weight_th}{the smallest value for weights in wls. Default to NULL.}

\item{plot_cv}{whether to plot error curves in cross validation.}
}
\value{
A list of K cell-type-specific co-expression matrices.
}
\description{
This function implements CSNet (estimation of Cell-type-Specific gene
co-expression Networks using bulk gene expression data) from the paper
DOI:10.1080/01621459.2023.2297467
}
