% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{generateData}
\alias{generateData}
\title{Generate simulation data for dynamic network inference}
\usage{
generateData(n0, p0, a, s, percent)
}
\arguments{
\item{n0}{number of samples}

\item{p0}{number of nodes}

\item{a}{range of random effects for nodes}

\item{s}{variance for random noise}

\item{percent}{proportion of non-zero coefficient}
}
\value{
simulated time t0, random effect a0, edge-edge coefficient G0 and F0, and data matrix X0
}
\description{
Generate simulation data for dynamic network inference
}
