% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_ls.R
\name{mom_ls}
\alias{mom_ls}
\title{Least squares for moment-based regressions}
\usage{
mom_ls(P, X, methods = list(var = "nnls", covar = "wls"), weight_th = NULL)
}
\arguments{
\item{P}{matrix of cell type proportions. n samples by K cell types.}

\item{X}{matrix of bulk gene expression. n samples by p genes.}

\item{methods}{least squares method to estimate variance and covariance.
Default to list(var='nnls', covar = 'wls'), i.e. non-negative least squares
for estimating variance and weighted least squares for estimating covariance.}

\item{weight_th}{the smallest value for weights in wls. Default to NULL.}
}
\value{
A list of K cell-type-specific variance-covariance matrices.
}
\description{
This function implements least squares (either ordinary least squares or
iteratively re-weighted least squares) for moment-based regressions to
estimate cell-type-specific covariances
}
