% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcRun.R
\name{runMCMC}
\alias{runMCMC}
\title{Graph structure and pseudotime inference by Markov chain Monte Carlo (MCMC)}
\usage{
runMCMC(X, d0, maxit, burnin, interval)
}
\arguments{
\item{X}{data matrix}

\item{d0}{number of basis for b-spline}

\item{maxit}{maximum number of iterations in MCMC}

\item{burnin}{discard samples before burnin period}

\item{interval}{retain samples every interval}
}
\value{
estimated graph structure G and pseudotime t
}
\description{
Graph structure and pseudotime inference by Markov chain Monte Carlo (MCMC)
}
