% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data_list}
\alias{sim_data_list}
\title{Simulated bulk gene expression and cell type proportion data}
\format{
\subsection{\code{who}}{

A list with three elements:
\describe{
\item{X}{matrix of bulk gene expression, n samples by p genes}
\item{P}{matrix of cell type proportions, n samples by K cell types}
\item{ct_X}{list of two cell-type-specific gene expression matrices,
each with n samples by p genes}
...
}
}
}
\source{
\url{https://github.com/ChangSuBiostats/CSNet_analysis}
}
\usage{
sim_data_list
}
\description{
Simulated data to demonstrate the use of CSNet, following the simulation
examples of DOI:10.1080/01621459.2023.2297467
}
\keyword{datasets}
